\name{elapsedtime}
\alias{elapsedtime}
\title{Internal Function}

\description{
Internal function.
}

\usage{
  elapsedtime(ptm1, ptm2)
}

\arguments{
  \item{ptm1}{\code{\link{proc.time}} object.}
  \item{ptm2}{\code{\link{proc.time}} object.}
}

\value{
List giving the elapsed time between \code{ptm1} and \code{ptm2}.
}

\keyword{ts}
